+++ utf8_string.cpp
@@ -47,7 +47,7 @@ namespace Sass {
 
     // function that will return a normalized index, given a crazy one
     size_t normalize_index(int index, size_t len) {
-      long signed_len = len;
+      long signed_len = static_cast<long>(len);
       // assuming the index is 1-based
       // we are returning a 0-based index
       if (index > 0 && index <= signed_len) {
